Scroll to top feature added
