+++ poshak-ui/src/components/HeroSection/index.js
@@ -36,7 +36,13 @@ const  HeroSection = () => {
                     onMouseEnter={onHover} 
                     onMouseLeave={onHover}
                     primary='true'
-                    dark='true'>
+                    dark='true'
+                    smooth={true} 
+                    duration={500} 
+                    spy={true} 
+                    exact='true' 
+                    offset={-80}
+                    >
                         Get started {hover ? <ArrowForward />:<ArrowRight />}
                     </Button>
                 </HeroBtnWrapper>
